feat: Ability to capture incremental stderr (#10179)
Exposes a new `Option` `ErrorProgressStreams` in the Automation API
that allows `stderr` to be redirected to one or more `io.Writers`.

The option is available for `refresh`, `destroy`, `preview` and `update`
operations. This will allow an enduser to tap into the ongoing stream
for `stderr`. Until now only concatenated `string` of outputs containing
all events was available in the returned result of an operation.