make python provider.py more tolerant of unexpected command line args (#10223)
When running Component Providers, command line flags are passed. When debugging using --logflow --logtostderr -v=5, -v=5 is passed to argparse. argparse doesn't expect this and crashes as it is not in the list of parameters the program expects.

This change makes the program more fault tolerant around unexpected parameters being passed in by callers.

Fixes #10208