Retry plugin downloads (#10248)
* Retry plugin downloads

We occasionally see network issues (most notably "tls: bad record MAC")
during plugin downloads. When grabbing plugins for codegen we already
had some retry logic. This shares that retry logic to the pulumi plugin install command and the engine.

Fixes #9824

Note, plumbing in the progress bar and support for printing about
retries is a little bit odd but we're going to have to totally rethink
this interface at some point when the engine gets put behind a gRPC
interface and all the display logic has to be ripped out. So it works
good enough for now.

* Add to CHANGELOG

* lint

* Finish comment