Work with all kinds of branch names (#10285)
* Generalise git branch test

In #10118 the git checkout was adapted so it would work with simple
branch names (i.e., "main" rather than "refs/heads/main"). However, it
still won't work with a branch that's not the default branch -- this
test demonstrates so.

* Support all simple and full branch names

If you want to have a non-default branch checked out, you need to fetch
it _and_ make a local reference for it. If you don't giv e a specific
reference, *Repository.Clone(...) updates remote heads, but does not
create local refs.

So, to support the varieties of branch names that people might use, the
provided branch name needs to be interpreted, and used for the
*Repository.Clone(...).

* Support cloning at a tag

Although the field is .Branch, we might expect people to ask for a tag
too. It's probably possible, with some back and forth, to disambiguate
between a simple branch name and a simple tag name; but, for the sake of
simplicity, require people to give `refs/tags/tag`.

* Queue changelog entry

* Check more cases of supplied branch name

 - ... including things that should result in an error
 - verify that the remote named in a remote ref is "origin"; anything
   else is surely a mistake

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>