Prototype of moving engine interface to gRPC
This changes the "about" command in pkg/cmd to simply be a formatting
shim, all the core logic of looking up backends/plugins/stacks etc is
now done by a new interface that cmd calls via grpc.

There's also a new "pulumi host-engine" hidden command to get the cli to
simply start up and print the address of the new grpc interface.

About was a fairly simple command to move over, it doesn't do any
explict logging or interactivity and is ok to block until all results
are collected and returned. But it lays the skeleton framework for the
new interface such that we can start iterating on this for other
commands, especially the new pulumi-deploy commands.