Fix regression validating project files with `refresh: always` set
We recently started validating the `Pulumi.yaml` file against its schema, leading to the following error when the `refresh: always` option is specified:

```
error: could not load current project: 1 error occurred:
	* #/options/refresh: expected boolean, but got string
```

The problem is the feature is expecting a string value of `"always"` but the schema for the project file is incorrectly specified as `boolean`.

This change fixes the schema, which avoids the error. A separate change will fix the documentation at https://www.pulumi.com/docs/reference/pulumi-yaml/#options-options