Fix update plans with dependent replacements
We weren't correctly handling the case where a resource was marked for
deletion due to one of it's dependencies being deleted. We would add an
entry to it's "Ops" list, but then overwrite that "Ops" list when we
came to generate the recreation step.

Fixes https://github.com/pulumi/pulumi/issues/10924