[auto] Fix hangs for remote previews
Retrieving engine events programmatically through automation api would result in the operation seemingly hanging for remote operations. This was particularly noticeable during remote previews, because preview in automation API programmatically inspects engine events to look for the summary event (which it doesn't do for up/refresh/destroy).

The problem is due to the lack of a terminating `CancelEvent`. Normally, for local operations, the event stream is always terminated by a `CancelEvent` which is the sentinel value used to indicate the operation is complete. However, when sending events to the service, the `CancelEvent` is not sent. So when we're retrieving events from the service for a remote operation, we weren't getting back a `CancelEvent`.

The fix is to send a synthesized `CancelEvent` after all the events have been read from the service (similar to what is done by the `pulumi replay-events` hidden debug command).