[cli] Reflow the interactive display as needed.
The interactive display currently has the unfortunate property that as
the width of the tree table grows, more and more content is lost off of
the right side of the screen. These changes add support for reflowing
the contents of the table as needed. Content is reflowed when the width
of the first N-1 columns of the N-column wide table exceed the width of
the terminal. We do not include the last column in the measurement
because the width of its contents vary wildly, whereas the width of the
other columns is essentially constant (or only increases).

When the terminal is wide enough, each node is layed out into a single
row of the form:

    op | indent + type | name | status | info

If the terminal is too narrow for the content of the first four columns,
each node is layed out into multiple rows:

    op | indent + type | status
       | name
       | info

Concrete examples are given below.

Wide display
------------

     Type                      Name                  Status            Info
     pulumi:pulumi:Stack       aws-ts-webserver-dev
     ├─ aws:ec2:SecurityGroup  web-secgrp
 +-  └─ aws:ec2:Instance       web-server-www        replaced (0.01s)  [diff: ~userData]

Outputs:
  ~ publicHostName: "ec2-34-211-56-110.us-west-2.compute.amazonaws.com" => undefined
  ~ publicIp      : "34.211.56.110" => undefined

Resources:
    +-1 replaced
    2 unchanged

Duration: 12s

Narrow display
--------------

     Type                      Status
     pulumi:pulumi:Stack
     aws-ts-webserver-dev
     ├─ aws:ec2:SecurityGroup
     │  web-secgrp
 +-  └─ aws:ec2:Instance       replaced (0.01s)
        web-server-www
        [diff: ~userData]

Outputs:
  ~ publicHostName: "ec2-34-211-56-110.us-west-2.compute.amazonaws.com" => undefined
  ~ publicIp      : "34.211.56.110" => undefined

Resources:
    +-1 replaced
    2 unchanged

Duration: 12s