[cli] Add a newline in the refresh confirmation prompt
The lack of a newline causes the prompt to wrap in terminals that are
fewer than 120 or so columns wide. The wrapping confuses the survey
package's redraw, which causes the repeated prompts each time the survey
is redrawn (i.e. on every keypress). Removing the wrapping by adding a
newline avoids this issue.