Add JSONMarshal to go sdk
This is not as useful for Go as it is for other languages, but it's
consistent and it saves an Apply call.

Ideally we would handle nested Output types during marshalling, but to
do that correctly we need some way to get access to
per-call-to-JSONMarshal flags to note if we see any unknowns, or
secrets.

Go just (AFAIK) have any way to support doing that.

So we do the next best thing which is we at least error that nested
Outputs can't be marshalled. Note that the `Output` interface doesn't
embed the `Marshal` interface because there _may_ be users who have
wrote their own `Output` types and that would be a breaking change.