fix(sdk/python): Allow for duplicate output values in python programs
I used the node SDK as inspiration for this change - there were a few things
we were doing differently in python (specifically handling all cases in
the outer layer of `massage` rather than allowing for more fine-grained
control in a `massage_complex` function like in node.  We also take the stack
concept from the node SDK to resolve the immediate issue of duplicate outputs.