Don't return half-nil fat ptrs from `SyntaxNode()`
Passing a typed `nil` through a interface conversion, such as a
`*hclsyntax.UnaryOpExpr` through a `hclsyntax.Node` interface will
result in a half-nil fat pointer with the value part `nil`. Since most
of programgen assumes that `(model.Expression).SyntaxNode().Range()` is
valid, a `syntax.None` is passed instead of a `nil`.