golangci-lint: Opt into nolintlint
This opts the project into [nolintlint],
which finds incorrect usages of the `//nolint` directive.
If used incorrectly, these can show up the project's API Reference,
which looks a bit shoddy.

   [nolintlint]: https://github.com/golangci/golangci-lint/blob/master/pkg/golinters/nolintlint/README.md

Note that this configures the linter to ignore unused directives.
We have a bunch of those across the project,
and some of them are in generated or templated files.

In addition to enabling the linter,
this fixes a couple other nolint directives
that were missed in #11791.

Depends on #11791
Relates to #11785