cli/update: Move experimental option to bottom
For the options presented in the 'Do you want to perform this update?'
dialog, move the experimental option to the bottom.

Before:

```
Do you want to perform this update?  [Use arrows to move, type to filter]
  [experimental] yes, using Update Plans (https://pulumi.com/updateplans)
  yes
> no
  details
```

Now:

```
Do you want to perform this update?  [Use arrows to move, type to filter]
  yes
> no
  details
  [experimental] yes, using Update Plans (https://pulumi.com/updateplans)
```

Resolves #11800