pkg/codegen/docs: Fix data race in tests
Tests in pkg/codegen/docs all run in parallel,
but many of them call initTestPackageSpec,
which modifies the global testPackageSpec variable.
So these tests end up stomping on each other.
`go test -race` fails in this directory.

To address the data race,
this change deletes the global variable entirely,
and replaces initTestPackageSpec with newTestPackageSpec.
newTestPackageSpec returns a new copy of the package spec
which the tests can use individually.

Resolves #11790