sdk/go/generate: Replace incorrect TrimRight with TrimSuffix
Code in go/pulumi/generate that determines the destination file path
based on the name of the template
incorrectly used TrimRight instead of TrimSuffix.

TrimRight removes all instances
of *any* of the provided characters (referred to as the cutset)
from the provided string
instead of removing just a fixed suffix.
So if we tried to generate say, a .tmpl file,
it would have stripped that from the destination file path too.

This switches to TrimSuffix, which is what was intended here,
and adds a test to verify the desired behavior.

Issue caught by staticcheck:

```
go/pulumi/generate/main.go:343:54: SA1024: cutset contains duplicate characters (staticcheck)
```

Refs #11808