ApplyT, ApplyTWithContext: De-dupe applier validation
ApplyT duplicates the applier validation logic used by ApplyTWithContext
with the exception that it expects one parameter instead of two.
It then wraps the applier in a wrapper function which ignores the context,
and passes it on to ApplyTWithContext,
which validates the function *again*.

This change de-dupes the validation logic
by introducing an internal "applier" object.
This object represents validated and normalized versions
of user-supplied appliers.

Given an applier object, ApplyTWithContext can use applier.Call
and always pass in a context, and always expect an error return.

    o, err := ap.Call(ctx, i)

This change is made in service to #11783 and #11784