sdk/go: Fix TestDependsOnInputs data race
TestDependsOnInputs modified the `monitor` attached to the context
while the context was still using it.
This caused a data race:

```
WARNING: DATA RACE
Read at 0x00c00041d8d8 by goroutine 4280:
  github.com/pulumi/pulumi/sdk/v3/go/pulumi.(*interceptingResourceMonitor).RegisterResource()
      [..]/pulumi/sdk/go/pulumi/resource_test.go:708 +0x44
  github.com/pulumi/pulumi/sdk/v3/go/pulumi.(*Context).registerResource.func1()
      [..]/pulumi/sdk/go/pulumi/context.go:854 +0xd28

Previous write at 0x00c00041d8d8 by goroutine 2880:
  github.com/pulumi/pulumi/sdk/v3/go/pulumi.newInterceptingResourceMonitor()
      [..]/pulumi/sdk/go/pulumi/resource_test.go:671 +0xac
```

Replacing the ResourceMonitorClient appears to be the only way
to get the information that the dependency tracker needs,
but replacing the client after Context creation is a data race.

To address this issue, introduce a new internal RunOption
that allows installing a ResourceMonitorClient middleware
at creation.

    WrapResourceMonitorClient(
        func(ResourceMonitorClient) ResourceMonitorClient,
    ) RunOpton

With this option present, we can decorate the client
with the dependency tracker
when the context gets created rather than afterwards.

Refs #10092