ci: Wait to run tests till after linting
The test matrix takes up a lot of workers.
If lint is failing, this will all have to re-run anyway.

We can free up resources for other PRs
by making running of tests conditional on linting.

A disadvantage of this is that it'll take
a couple minutes longer to start running tests.
Linting takes 2-4 minutes.