cmd/stack output: Pull out core logic
Pull core logic of `pulumi stack output`
into a separate `stackOutputCmd` struct
rather than having it all inside an anonymous function.
That also turns state captured in the closure
into normal fields on the struct.

This contains no logic changes.

This is a step towards making this logic more testable
in service to #2632.