cmd: Reduce noise from offerNew, setCurrent bool
There are a few helper functions in util.go
that accept two booleans: offerNew and setCurrent.
These functions are called *quite* often
and callers always have to pass these two naked booleans:

    requireStack(ctx, stack, false, opts, true)

To reduce confusion, these booleans have to be annotated with a comment.

    requireStack(ctx, stack, false, opts, true /* setCurrent */)

Given how pervasive these two are,
this change turns them into a set of bit flags:

    stackLoadOnly
    stackOfferNew
    stackSetCurrent

Callers pass in a bitwise OR of these as needed:

           Before                   Now
    offerNew  setCurrent       stackLoadOption
    ------------------------------------------
      false     false           stackLoadOnly
      true      false           stackOfferNew
      false     true            stackSetCurrent
      true      true    stackOfferNew|stackSetCurrent

This reduces risk of passing the booleans out of order,
and improves readability at those call sites.