cmd/stack output: Print to an io.Writer
The implementation of `pulumi stack output` prints everything
to a global `io.Writer`.
This makes testing its output a bit hacky and unparalellizable
because we have to hijack `os.Stdout` to capture the output.

This changes it the command implementation
and all the functions it relies on
to accept an `io.Writer` as an argument.

This allows us to delete the os.Stdout hijacking hack
that was introduced in #11952,
and have those tests run in parallel, writing to an in-memory buffer.