Move test into NodeSDK
tests/integration/ is not set up run multi-directory programs.
This integration test is better suited for the Node SDK, since
it tests a NodeJS feature. The test is intended to ensure the Node
language runtime locates a package.json file in a parent directory
of the Pulumi program, a pattern used in monorepos. The language runtime
is aimed at a directory without an index.js file; it must find the
package.json file in the parent directory, read it's "main" property
to find the program entrypoint, and then execute that entrypoint
without error. This validates that the runtime finds package.json files
higher in the directory hierarchy.