feat: enable simpler Go programtest replacements
This change to ProgramTest allows Go module replacements to omit the major
version number, simplifying the process to release a provider with a new major
version.

When the major version is incremented in a PR, the examples fail to compile if
they reference the incorrect SDK version.

This enables the examples to be written to target a versionless `/sdk`, and have
a Dependencies line such as:

```go
	integration.ProgramTestOptions{
		Dependencies: []string{
			"github.com/pulumi/pulumi-keycloak/sdk=../sdk",
		},
	}
```

This versionless import mapping means that examples run regardless of the major
version of the provider, and updates to major versions in PRs will require only
one PR.

A provider PR proving out the implementation was created against this commit
here:
- https://github.com/pulumi/pulumi-keycloak/pull/165