sdk/go: Support looking inside []InvokeOption
This is the InvokeOption analog for NewResourceOptions added in #12124.

However, in contrast to ResourceOptions, this just exports the existing
invokeOptions struct.
The reason for this is that right now, the public API for InvokeOptions
is the same as the internal representation,
so we don't need the separation-- for now.
To keep future edits aware of this distinction,
I've included a loud comment about what to do
when they do inevitably diverge.

Refs #11698