sdk/go: Don't store DependsOn in a lossy form
The `DependsOn` and `DependsOnInputs` resource options
store their captured information on the `resourceOptions` struct
in a lossy format: they store function references.

This makes it impossible to go back to the original lists of resources
or resource array inputs for use cases like #11698.

As a step towards making this possible,
replace the stored closures with interfaces.
The implementations are a drop-in replacement for the prior behavior.

Refs #11698