test/integration: Don't panic during setup
Currently, component setup panics if there's an error.
This isn't great because when it panics,
it fails to notify the outer goroutine that's waiting for setup,
which leaves that goroutine waiting for 10 minutes gefore giving up.

The net effect of this is that when setup fails for a test,
it takes 10 minutes to kill the test
even if setup failed within seconds.

Fix this by using testify and logging errors right away.