Remove our use of filelocks in component tests
This didn't need to be file locked, we don't run test processes in
parallel, it's sufficently covered by the YarnInstallMutex we take when
entering the function.

That mutex is because it's not safe to run yarn concurrently across the
machine, but again because we're running all the tests in one go process
a normal `sync.Mutex` suffices for that as well.